requirements.txt
gunicorn==20.0.4
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Sockets==0.2.1
protobuf==3.6.1
scikit-learn==0.22.2
python-dotenv==0.12.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.12.0
gevent==20.5.0
redis==3.5.2
requests==2.24.0